refac: file structure is now separated into 2 segments
