+++ src/components/Map/Map.js
@@ -1,9 +1,9 @@
-import LoopCastles from "./LoopCastleFlagMarks";
+import LoopCastleFlagMarks from "./LoopCastleFlagMarks";
 
 function Map() {
   return (
     <div>
-      <LoopCastles />
+      <LoopCastleFlagMarks />
       <img src="images/PE_map.jpg" className="mw-100" />
     </div>
   );
