+++ src/components/Map/ResourceMarks/ResourceMarks.js
