+++ src/components/Menu/ResourceList/ResourceIcon/ResourceIcon.module.css
